Make use of GHC 9.2
This commit should really be split up.

Note that there's one breaking change in an internal module - we now use `id` for a field name in `Product`.

Note that the `msgResWitness` change actually could be performed without `ImpredicativeTypes` if we switched the argument order. Or even used `$` instead of `BlockArguments`, due to the special typing rule for `$`, which has existed as a workaround for the lack of `ImpredicativeTypes`. (it also doesn't work as hoped - see where I've attempted to use it)